+++ integrations/anthropic/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [integrations/anthropic-v4.1.0] - 2025-08-05
+
+### 🚀 Features
+
+- Adopt new `StreamingChunk` in Anthropic (#2096)
+
 ## [integrations/anthropic-v3.1.0] - 2025-07-04
 
 ### 🚀 Features